fix(StringCalculator): 큰 수 입력 테스트 실패 문제 해결
* 문자열 계산기 특성 상 입력 될 수 있는 큰 수 입력 문제를 해결
  * Before: Fail, After: Success
* Long Type 에서 BigInteger 타입으로 변경

+++ src/main/java/calculator/v1/Controller.java
@@ -1,5 +1,7 @@
 package calculator.v1;
 
+import java.math.BigInteger;
+
 public class Controller {
 
     private final View view;
@@ -12,7 +14,7 @@ public class Controller {
         view.showStringInput();
         String expression = view.readStringExpression();
         StringCalculator stringCalculator = new StringCalculator(expression);
-        long result = stringCalculator.sum();
+        BigInteger result = stringCalculator.sum();
         view.showResult(result);
     }
 }
